Arreglé errores en mis funciones

+++ TrabajoPráctico_1/src/OperacionesAritmeticas.h
@@ -1,7 +1,7 @@
-int Sumar(float, float); //Función que utilizo para sumar dos números.
+float Sumar(float, float); //Función que utilizo para sumar dos números.
 
-int Restar(float, float); //Función que utilizo para restar dos números.
+float Restar(float, float); //Función que utilizo para restar dos números.
 
-int Multiplicar(float, float); //Función para multiplicar dos números.
+float Multiplicar(float, float); //Función para multiplicar dos números.
 
-int Dividir(float, float); // Función para dividir entre dos números.
+float Dividir(float, float); // Función para dividir entre dos números.
